corrections

+++ binx/setup/setup-system.sh
@@ -134,44 +134,54 @@ function setup_git_repos()
     sudo mkdir -p $SITEDIR/.repos
     sudo chown ${USER}.${USER} -R $SITEDIR/.repos
 
-
-    if [ ! -d $SITEDIR/.repos/git/system/system ]
+    if [ ! -d $SITEDIR/.repos/git ]
     then
-	mkdir -p $SITEDIR/.repos/git/system/
-	git clone git@github.com:sharad/system.git $SITEDIR/.repos/git/system/system
+        git clone --recursive git@github.com:sharad/usrlocalorg.git $SITEDIR/.repos/git
+    else
+        git -C $SITEDIR/.repos/git submodule update --remote
     fi
 
-    if [ ! -L $SITEDIR/.system ]
+    if false
     then
-	      rm -rf $SITEDIR/.system
-	      sudo ln -sf .repos/git/user/system/system $SITEDIR/.system
-    fi
 
-    if [ ! -d $SITEDIR/.repos/git/packages/misc/browser/conkeror ]
-    then
-	mkdir -p $SITEDIR/.repos/git/packages/misc/browser
-	git clone git://repo.or.cz/conkeror.git $SITEDIR/.repos/git/packages/misc/browser/conkeror
-    fi
+        if [ ! -d $SITEDIR/.repos/git/system/system ]
+        then
+	    mkdir -p $SITEDIR/.repos/git/system/
+	    git clone git@github.com:sharad/system.git $SITEDIR/.repos/git/system/system
+        fi
 
-    if [ ! -d $SITEDIR/.repos/git/packages/common-lisp/source/sharad/in.net.sharad.utils ]
-    then
-	mkdir -p $SITEDIR/.repos/git/packages/common-lisp/source/sharad
-	git clone git@bitbucket.org:sh4r4d/in.net.sharad.utils $SITEDIR/.repos/git/packages/common-lisp/source/sharad/in.net.sharad.utils
-    fi
+        if [ ! -L $SITEDIR/.system ]
+        then
+	    rm -rf $SITEDIR/.system
+	    sudo ln -sf .repos/git/user/system/system $SITEDIR/.system
+        fi
 
-    if [ ! -d $SITEDIR/.repos/git/packages/common-lisp/source/sharad/pa ]
-    then
-	mkdir -p $SITEDIR/.repos/git/packages/common-lisp/source/sharad
-        git clone git@bitbucket.org:sh4r4d/pa $SITEDIR/.repos/git/packages/common-lisp/source/sharad/pa
-    fi
+        if [ ! -d $SITEDIR/.repos/git/packages/misc/browser/conkeror ]
+        then
+	    mkdir -p $SITEDIR/.repos/git/packages/misc/browser
+	    git clone git://repo.or.cz/conkeror.git $SITEDIR/.repos/git/packages/misc/browser/conkeror
+        fi
 
-    if [ ! -d $SITEDIR/.repos/git/packages/common-lisp/source/sharad/stumpwm ]
-    then
-	mkdir -p $SITEDIR/.repos/git/packages/common-lisp/source/sharad
-        git clone https://sharad@github.com/sharad/stumpwm.git $SITEDIR/.repos/git/packages/common-lisp/source/sharad/stumpwm
-        cd $SITEDIR/.repos/git/packages/common-lisp/source/sharad/stumpwm
-        git checkout pa-point-timeout
-        cd -
+        if [ ! -d $SITEDIR/.repos/git/packages/common-lisp/source/sharad/in.net.sharad.utils ]
+        then
+	    mkdir -p $SITEDIR/.repos/git/packages/common-lisp/source/sharad
+	    git clone git@bitbucket.org:sh4r4d/in.net.sharad.utils $SITEDIR/.repos/git/packages/common-lisp/source/sharad/in.net.sharad.utils
+        fi
+
+        if [ ! -d $SITEDIR/.repos/git/packages/common-lisp/source/sharad/pa ]
+        then
+	    mkdir -p $SITEDIR/.repos/git/packages/common-lisp/source/sharad
+            git clone git@bitbucket.org:sh4r4d/pa $SITEDIR/.repos/git/packages/common-lisp/source/sharad/pa
+        fi
+
+        if [ ! -d $SITEDIR/.repos/git/packages/common-lisp/source/sharad/stumpwm ]
+        then
+	    mkdir -p $SITEDIR/.repos/git/packages/common-lisp/source/sharad
+            git clone https://sharad@github.com/sharad/stumpwm.git $SITEDIR/.repos/git/packages/common-lisp/source/sharad/stumpwm
+            cd $SITEDIR/.repos/git/packages/common-lisp/source/sharad/stumpwm
+            git checkout pa-point-timeout
+            cd -
+        fi
     fi
 
 }
